release-20250830-194850/requirements-prod.txt
# 本番環境で必要な依存関係
Flask>=2.0.0
flask-babel>=2.0.0
Flask-Login>=0.6.0
Flask-SQLAlchemy>=3.0.0
pillow>=9.0.0
PyMySQL>=1.0.0
click>=8.0.0
cryptography>=3.4.0
pyotp>=2.6.0
python-dotenv>=0.19.0
qrcode>=7.3.0
requests>=2.28.0
celery>=5.2.0
redis>=4.3.0
PyJWT>=2.4.0
Markdown>=3.4.0
python-dateutil>=2.8.0
MarkupSafe>=2.1.0
gunicorn>=20.1.0